backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Pydantic AI with ARCHON
pydantic-ai==0.0.14
pydantic==2.9.2

# AI/ML Libraries
openai==1.54.3
anthropic==0.39.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0

# Database
supabase==2.9.1
psycopg2-binary==2.9.10
pgvector==0.3.5

# Vector Embeddings
sentence-transformers==3.2.1
numpy==2.1.3

# Utilities
httpx==0.27.2
aiohttp==3.11.0
tenacity==9.0.0
